dice-updater/requirements.txt
selenium==4.8.2
webdriver-manager==3.8.5
PyYAML~=6.0
schedule==1.1.0
